crawler/requirements.txt
httpx>=0.27.0
playwright>=1.45.0
anthropic>=0.30.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
lxml>=5.2.0
